Starting to refactor to create the actors every time
