+++ src/main/scala/is/kow/scalatratrackerapp/actors/commands/CommandResponder.scala
@@ -1,0 +1,5 @@
+package is.kow.scalatratrackerapp.actors.commands
+
+trait CommandResponder {
+  val commandRegex:String
+}
