+++ src/main/scala/is/kow/scalatratrackerapp/actors/responders/TextResponder.scala
@@ -1,0 +1,7 @@
+package is.kow.scalatratrackerapp.actors.responders
+
+import scala.util.matching.Regex
+
+trait TextResponder {
+  val regexes: Seq[Regex]
+}
